combined navigation compact into navigation fluent
