+++ src/Wpf.Ui.Demo/Views/Windows/ExperimentalWindow.xaml
@@ -24,18 +24,20 @@
             <ColumnDefinition Width="*" />
         </Grid.ColumnDefinitions>
 
-        <ui:NavigationCompact
+        <ui:NavigationFluent
             x:Name="RootNavigation"
             Grid.Column="0"
             Frame="{Binding ElementName=RootFrame}"
+            IsExpandButtonVisible="True"
+            IsExpanded="False"
             Precache="True">
-            <ui:NavigationCompact.Footer>
+            <ui:NavigationFluent.Footer>
                 <ui:NavigationItem
                     Click="NavigationButtonTheme_OnClick"
                     Content="Theme"
                     Icon="DarkTheme24" />
-            </ui:NavigationCompact.Footer>
-        </ui:NavigationCompact>
+            </ui:NavigationFluent.Footer>
+        </ui:NavigationFluent>
 
         <ui:TitleBar
             Title="WPF UI - Experimental Page"
